requirements.txt
flask==2.0.1
tensorflow==2.8.0
numpy==1.21.5
opencv-python==4.5.5.64
Pillow==9.0.1
werkzeug==2.0.2
scikit-learn==1.0.2
matplotlib==3.5.1 